+++ api/src/User/Service/UserManager.php
@@ -1458,6 +1458,8 @@ class UserManager
                     $bankaccount->setIban($iban);
                     $bankaccount->setBic($bic);
                     
+                    $bankaccount->setRefusalReason($paymentProfile->getRefusalReason());
+
                     $bankAccounts[] = $bankaccount;
                 }
             }